+++ RELEASES.md
@@ -18,6 +18,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template: fixed issue with QualifiedSpecialization in Instantiation
 * prov-model, prov-template, prov-dot: support for QualifiedHadMember
 * prov-template-compiler: generation of OutputProcessor interface
+* prov-toolbox-plugin: renamed goals, validate-catalogue and compile-catalogue
 
  
 `2.2.1`